workflow/requirements.txt
alembic==1.3.2
amqp==5.1.1
async-timeout==4.0.2
attrs==22.1.0
billiard==3.6.4.0
celery==5.2.6
celery-director==0.6.0
certifi==2022.6.15
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
Deprecated==1.2.13
environs==6.1.0
Flask==2.1.1
Flask-HTTPAuth==4.5.0
Flask-json-schema==0.0.5
Flask-Migrate==3.1.0
Flask-SQLAlchemy==2.5.1
flower==1.1.0
gunicorn==20.0.4
humanize==4.2.3
idna==3.3
importlib-metadata==4.12.0
importlib-resources==5.9.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsonschema==4.9.1
kombu==5.2.4
Mako==1.2.1
MarkupSafe==2.1.1
marshmallow==3.17.0
mysql-connector-python==8.0.21
packaging==21.3
pkgutil-resolve-name==1.3.10
pluginbase==1.0.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
protobuf==4.21.4
psycopg2-binary==2.9.3
pyparsing==3.0.9
pyrsistent==0.18.1
python-dateutil==2.8.2
python-dotenv==0.20.0
python-editor==1.0.4
pytz==2022.1
PyYAML==5.4.1
redis==4.2.2
requests==2.28.1
sentry-sdk==0.16.3
six==1.16.0
SQLAlchemy==1.3.11
SQLAlchemy-Utils==0.35.0
terminaltables==3.1.0
tornado==6.2
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.1.1
wrapt==1.14.1
zipp==3.8.1
